# openapi.yaml
openapi: 3.0.3
info:
  title: PPI API
  description: API para gestion de órdenes de inversión - PPI
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local Docker environment
  - url: https://localhost:5001
    description: Local development server
security:
  - BearerAuth: []
paths:
...
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: Autentica al usuario y genera un token JWT
      description: Se puede usar cualquier usuario/contraseña ya que no hay tabla de usuarios en la base de datos.
      security: []  
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequestDto'
            example:
              username: "demo"
              password: "demo"
      responses:
        '200':
          description: Token JWT para autorización
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponseDto'
        '400':
          description: La información recibida no es válida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
  /api/auth/verify:
    get:
      tags:
        - Authentication
      summary: Endpoint de prueba para verificar que la autenticación JWT funciona
      description: Información del usuario autenticado
      responses:
        '200':
          description: Token válido
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Token válido"
                  username:
                    type: string
                    example: "demo"
                  claims:
                    type: array
                    items:
                      type: object
                      properties:
                        type:
                          type: string
                        value:
                          type: string
        '401':
          description: No autorizado

  # Asset Controller
...
  /api/asset/{id}:
    get:
      tags:
        - Assets
      summary: Obtiene un activo específico por su Id
      description: Activo encontrado o NotFound si no existe
      parameters:
        - name: id
          in: path
          required: true
          description: Id único del activo
          schema:
            type: integer
      responses:
        '200':
          description: Activo encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetResponse'
        '404':
          description: Activo no encontrado
  /api/asset/GetAssetsByAssetTypeId/{assetTypeId}:
    get:
      tags:
        - Assets
      summary: Obtiene todos los activos que pertenecen a un tipo específico
      description: |
        Lista de activos del tipo especificado o NotFound si no existen.
        Tipos de activo: 1 = Acción, 2 = Bonos, 3 = FCI
      parameters:
        - name: assetTypeId
          in: path
          required: true
          description: Id del tipo de activo
          schema:
            type: integer
            enum: [1, 2, 3]
      responses:
        '200':
          description: Lista de activos del tipo especificado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetListResponse'
        '404':
          description: No se encontraron activos para el tipo especificado

  # AssetType Controller
...
  /api/assettype/{id}:
    get:
      tags:
        - Asset Types
      summary: Obtiene un tipo de activo específico por su Id
      description: |
        Tipo de activo encontrado o NotFound si no existe.
        Tipos: 1 = Acción, 2 = Bonos, 3 = FCI
      parameters:
        - name: id
          in: path
          required: true
          description: Id único del tipo de activo
          schema:
            type: integer
            enum: [1, 2, 3]
      responses:
        '200':
          description: Tipo de activo encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetTypeResponse'
        '404':
          description: Tipo de activo no encontrado

  # OperationType Controller
...
  /api/operationtype/{id}:
    get:
      tags:
        - Operation Types
      summary: Obtiene un tipo de operación específico por su Id
      description: |
        Tipo de operación encontrado o NotFound si no existe.
        Tipos: C = Compra, V = Venta
      parameters:
        - name: id
          in: path
          required: true
          description: Id único del tipo de operación
          schema:
            type: string
            enum: ["C", "V"]
      responses:
        '200':
          description: Tipo de operación encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationTypeResponse'
        '404':
          description: Tipo de operación no encontrado

  # OrderState Controller
...
  /api/order/asset/{assetId}:
    get:
      tags:
        - Orders
      summary: Obtiene todas las órdenes asociadas a un activo específico
      description: Lista de órdenes del activo especificado
      parameters:
        - name: assetId
          in: path
          required: true
          description: Id del activo
          schema:
            type: integer
      responses:
        '200':
          description: Lista de órdenes del activo especificado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderListResponse'
...
  /api/order/operation/{operationTypeId}:
    get:
      tags:
        - Orders
      summary: Obtiene todas las órdenes que tienen un tipo de operación específico
      description: |
        Lista de órdenes con el tipo de operación especificado.
        Tipos: C = Compra, V = Venta
      parameters:
        - name: operationTypeId
          in: path
          required: true
          description: Id del tipo de operación
          schema:
            type: string
            enum: ["C", "V"]
      responses:
        '200':
          description: Lista de órdenes con el tipo de operación especificado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderListResponse'
  /api/order/assettype/{assetTypeId}:
    get:
      tags:
        - Orders
      summary: Obtiene todas las órdenes asociadas a un tipo de activo específico
      description: |
        Lista de órdenes del tipo de activo especificado.
        Tipos: 1 = Acción, 2 = Bono, 3 = FCI
      parameters:
        - name: assetTypeId
          in: path
          required: true
          description: Id del tipo de activo
          schema:
            type: integer
            enum: [1, 2, 3]
      responses:
        '200':
          description: Lista de órdenes del tipo de activo especificado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderListResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Authorization header usando el esquema Bearer. Ejemplo "Authorization Bearer {token}"
  schemas:
...
    LoginRequestDto:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          minLength: 3
          description: Nombre de usuario (mínimo 3 caracteres)
          example: "demo"
        password:
          type: string
          minLength: 4
          description: Contraseña (mínimo 4 caracteres)
          example: "demo"
    LoginResponseDto:
      type: object
      properties:
        token:
          type: string
          description: Token JWT
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        expires:
          type: string
          format: date-time
          description: Fecha y hora de expiración del token
          example: "2025-09-01T12:00:00Z"
        tokenType:
          type: string
          description: Tipo de token
          example: "Bearer"

    # Order DTOs
...
    ResultModel:
      type: object
      properties:
        data:
          oneOf:
            - type: object
            - type: array
          description: Datos de la respuesta
        ok:
          type: boolean
          description: Indica si la operación fue exitosa (true si no hay errores)
        message:
          type: string
          description: Mensaje de la operación
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ErrorModel'
          description: Lista de errores si los hay
    ErrorModel:
      type: object
      properties:
        message:
          type: string
          description: Mensaje del error
        errorCode:
          type: integer
          enum: [0, 1, 2]
          description: Código del error (0=InputDataError, 1=DatabaseError, 2=InternalError)

    # Domain Models
    AssetModel:
      type: object
      properties:
        id:
          type: integer
          description: Id único del activo
          example: 1
        ticker:
          type: string
          description: Símbolo del activo
          example: "AAPL"
        name:
          type: string
          description: Nombre del activo
          example: "Apple Inc."
        assetType:
          type: integer
          description: Tipo de activo (1=Acción, 2=Bonos, 3=FCI)
          enum: [1, 2, 3]
          example: 1
        unitPrice:
          type: number
          format: decimal
          description: Precio unitario del activo
          example: 150.25
    AssetTypeModel:
      type: object
      properties:
        id:
          type: integer
          description: Id único del tipo de activo
          example: 1
        description:
          type: string
          description: Descripción del tipo de activo
          example: "Acción"
    OperationTypeModel:
      type: object
      properties:
        id:
          type: string
          description: Id del tipo de operación
          example: "C"
        description:
          type: string
          description: Descripción del tipo de operación
          example: "Compra"
...
    OrderDetailedModel:
      type: object
      properties:
        id:
          type: integer
          description: Id único de la orden
          example: 1
        accountId:
          type: integer
          description: Id de la cuenta
          example: 1
        assetId:
          type: integer
          description: Id del activo
          example: 1
        quantity:
          type: integer
          format: int64
          description: Cantidad de activos
          example: 100
        price:
          type: number
          format: decimal
          description: Precio del activo
          example: 150.25
        operationTypeId:
          type: string
          description: Tipo de operación
          example: "C"
        statusId:
          type: integer
          description: Estado de la orden
          example: 0
        totalAmount:
          type: number
          format: decimal
          description: Monto total de la orden
          example: 15025.00
        assetName:
          type: string
          description: Nombre del activo
          example: "Apple Inc."
        assetType:
          type: integer
          description: Tipo de activo
          example: 1
        assetTypeDescription:
          type: string
          description: Descripción del tipo de activo
          example: "Acción"
        operationTypeDescription:
          type: string
          description: Descripción del tipo de operación
          example: "Compra"
        statusDescription:
          type: string
          description: Descripción del estado
          example: "En proceso"

    # Specific Response Types
    AssetResponse:
      allOf:
        - $ref: '#/components/schemas/ResultModel'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/AssetModel'
    AssetListResponse:
      allOf:
        - $ref: '#/components/schemas/ResultModel'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/AssetModel'
    AssetTypeResponse:
      allOf:
        - $ref: '#/components/schemas/ResultModel'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/AssetTypeModel'
...
    OperationTypeResponse:
      allOf:
        - $ref: '#/components/schemas/ResultModel'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/OperationTypeModel'
...
    OrderListResponse:
      allOf:
        - $ref: '#/components/schemas/ResultModel'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/OrderDetailedModel'
    ValidationError:
      type: object
      properties:
        success:
          type: integer
          example: 0
          description: Indica que la operación falló
        message:
          type: string
          example: "La información recibida no es válida. Por favor, revise los datos e intente nuevamente."
          description: Mensaje de error de validación
...